requirements.txt
vllm
transformers
torch
torchvision
Pillow
numpy
nltk
gradio
opencv-python
openai
anthropic
google-genai
python-dotenv
